Update delta to contain 'b' indicating literal byte next
Previous to this change, we were using '@' to indicate a reference chunk, but '@' is a valid character that could be part of the string itself. In those cases, we were parsing a '@' erroneously as a reference token. Now all byte literals follow a 'b', so we can differentiate between a token '@' and a byte '@'.